requirements.txt
asgiref==3.10.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.0.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
whitenoise==6.11.0
